pyproject.toml
[tool.poetry]
name = "hop-barley-shop"
version = "0.1.0"
description = "Интернет-магазин крафтового пива Hop & Barley"
authors = ["Artem <artem@email.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=5.2"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.0"
pillow = "10.2.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "0.27.0"
django-filter = "^25.2"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "4.2.0"
django-extensions = "3.2.3"
pytest = "7.4.4"
pytest-django = "4.7.0"
pytest-cov = "4.1.0"
factory-boy = "3.3.0"
flake8 = "7.0.0"
isort = "*"
mypy = "1.8.0"
django-stubs = "4.2.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"